Level 0: floor the recorded times, not the running clock (S162)
traverse_free_flow floored every link's exit and carried the floored value
forward as the next link's entry, so each link lost its sub-second
remainder: about half a second per link on average, about a minute on a
hundred-link trip, and 164 s against 165.99 s on the toy network's seven-link
route (levels 2-4 give 165 s). The clock is now kept exact and only what is
recorded is floored, which is what S88 says and what levels 2-4 do.

Costs nothing per link or vehicle. Level-0 trip times change by the lost
remainders (longer, the correct direction); the two tests that pinned the
old link-by-link flooring now assert the property, floor of the exact
cumulative time.

+++ crates/core-sim/tests/run.rs
@@ -83,14 +83,13 @@ fn car_owning_defaults() -> ClassDefaults {
     ClassDefaults::new().with_default("commuter", Ownership { car: true, ..Ownership::NONE })
 }
 
-/// The free-flow time of `a -> b -> c`, floored per link the same way
-/// `core-loading` does — the exact number a completed a-to-c trip's travel
-/// time must equal at weight 1.
+/// The free-flow time of `a -> b -> c`, floored once (S88; S161, F1) — the
+/// exact number a completed a-to-c trip's travel time must equal at weight 1.
 fn expected_a_to_c_seconds(network: &RoadNetwork) -> u32 {
     let ab = network.link_external_ids().typed_id_of::<LinkId>("ab").expect("known link");
     let bc = network.link_external_ids().typed_id_of::<LinkId>("bc").expect("known link");
     #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss, reason = "test assertion")]
-    let seconds = network.free_flow_time(ab).get() as u32 + network.free_flow_time(bc).get() as u32;
+    let seconds = (network.free_flow_time(ab).get() + network.free_flow_time(bc).get()) as u32;
     seconds
 }
 
@@ -148,8 +147,8 @@ fn a_car_trip_ignores_footways() {
     assert_eq!(result.completion.completed, 1);
     let id = |e: &str| network.link_external_ids().typed_id_of::<LinkId>(e).expect("link");
     #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss, reason = "test assertion")]
-    let by_road = network.free_flow_time(id("ab")).get() as u32
-        + network.free_flow_time(id("bc")).get() as u32;
+    let by_road =
+        (network.free_flow_time(id("ab")).get() + network.free_flow_time(id("bc")).get()) as u32;
     #[allow(clippy::float_cmp, reason = "both sides are small integer seconds, exactly")]
     {
         assert_eq!(result.total_travel_time.get(), f64::from(by_road), "the car takes ab, bc");
